Tính năng (lịch): Danh sách giao dịch chi tiết & số dư ròng hàng ngày Tính năng (biểu đồ): Thêm biểu đồ tròn, biểu đồ cột và biểu đồ đường với hiệu ứng động Sửa lỗi (giao dịch): Khắc phục tình trạng tranh chấp khi lưu dữ liệu Sửa lỗi (giao diện người dùng): Trạng thái điều hướng dưới cùng & tính toán tóm tắt hàng ngày Tái cấu trúc: Hợp nhất ngân sách vào màn hình chính & cải thiện tương tác danh sách

+++ app/src/main/java/com/nhattien/expensemanager/data/dao/TransactionDao.kt
@@ -24,5 +24,5 @@ interface TransactionDao {
     fun getTransactionsInRange(startDate: Long, endDate: Long): Flow<List<TransactionEntity>>
 
     @Query("SELECT * FROM transactions WHERE id = :id")
-    suspend fun getById(id: Int): TransactionEntity? // Sửa lại thành Int nếu ViewModel đang dùng Int, hoặc Long cho chuẩn
+    suspend fun getById(id: Long): TransactionEntity?
 }
